// src/app/interfaces/Track.ts
export interface Track {
  name: string; // Nombre de la canción
  artist: string; // Nombre del primer artista
  albumImage: string; // URL de la imagen del álbum
  popularity: number; // Popularidad de la canción
  previewUrl: string; // URL de pre-escucha de la canción
  spotifyUrl: string; // Enlace de Spotify para escuchar la canción
}